feat: reworked dotfiles

+++ fish/scripts/dockutils.fish
@@ -2,16 +2,22 @@
 
 set -l dockutils_desc "Utility functions for Docker container management"
 function dockutils --description $dockutils_desc
-    argparse 'h/help' 's/stop' 'c/clean' -- $argv or return
+    argparse 'h/help' 's/stop' 'c/clean' 'y/yes' -- $argv or return
 
     if set -ql _flag_help
         help-view \
-            --usage="dockutils [--stop|-s] [--clean|-c]" \
+            --usage="dockutils [--stop|-s] [--clean|-c] [--yes|-y]" \
             --description=$dockutils_desc \
+            --dependency="docker" \
             --opt="s/stop Stop all running Docker containers." \
-            --opt="c/clean Clean up all unused Docker data (containers, images, networks, build cache, volumes)."
+            --opt="c/clean Clean up all unused Docker data (containers, images, networks, build cache, volumes)." \
+            --opt="y/yes Skip confirmation for destructive cleanup." \
+            --example="dockutils --stop" \
+            --example="dockutils --clean --yes"
         return
     end
+
+    __require_command docker or return
 
     if set -ql _flag_stop
         __stop
@@ -19,11 +25,11 @@ function dockutils --description $dockutils_desc
     end
 
     if set -ql _flag_clean
-        __clean
+        __clean (set -q _flag_yes; and echo yes)
         return
     end
 
-    echo "Invalid option: $argv[1]. Use --help for usage information."
+    echo "Choose an action. Use --help for usage information."
     return 1
 end
 
@@ -44,8 +50,16 @@ function __stop
     echo "All running Docker containers stopped."
 end
 
-function __clean
+function __clean --argument-names confirmed
     echo "Cleaning up all unused Docker data (containers, images, networks, build cache, volumes)..."
+
+    if test "$confirmed" != yes
+        __confirm "This will remove unused Docker containers, images, networks, build cache, and volumes."
+        or begin
+            echo "Cancelled."
+            return 1
+        end
+    end
 
     docker system prune -af --volumes
     if test $status -ne 0
@@ -54,3 +68,8 @@ function __clean
     end
     echo "Docker system clean-up complete."
 end
+
+complete -c dockutils -f
+complete -c dockutils -s s -l stop -d "Stop all running Docker containers"
+complete -c dockutils -s c -l clean -d "Clean unused Docker data"
+complete -c dockutils -s y -l yes -d "Skip cleanup confirmation"
